pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-deploy-cli"
version = "0.1"
description = "A CLI tool for DevOps deployment"
requires-python = ">=3.9,<3.10"
dependencies = [
    "Click",
    "PyYAML",
    "ansible",
    "configparser",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0",
    "ruff>=0.9.4,<1.0",
]

[tool.setuptools]
packages = ["cli_commands", "cli_utils"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.ini"]